+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/RouteHandlers/DisallowNonParsableComplexTypesOnParameters.cs
@@ -64,7 +64,9 @@ public partial class RouteHandlerAnalyzer : DiagnosticAnalyzer
                 location
                 )) { continue; }
 
-            var routeParameterName = RouteParameterHelper.ResolveRouteParameterName(handlerDelegateParameter, wellKnownTypes);
+
+            var parameterSymbol = routeUsage.UsageContext.ResolvedParameters.FirstOrDefault(p => p.Symbol.Name == handlerDelegateParameter.Name);
+            var routeParameterName = parameterSymbol.RouteParameterName ?? handlerDelegateParameter.Name;
 
             // Match handler parameter against route parameters. If it is a route parameter it needs to be parsable/bindable in some fashion.
             if (routeUsage.RoutePattern.TryGetRouteParameter(routeParameterName, out var routeParameter))
